requirements.txt
Flask==2.1.2
Flask-Login==0.6.1
psycopg2-binary
gunicorn
